+++ src/impl/translators/to-string.ts
@@ -5,9 +5,22 @@
  */
 
 import { factory } from "@/decorator-factory";
-import { PropertyDecorator, ReviserMessage } from "@inter/decorator";
+import { PropertyDecorator, ReviserMessage, ReviserDecoratorOptions } from "@inter/decorator";
+import { isRequired, getPrimitiveType, PrimitiveTypes } from "@impl/utils";
 
-function decorator(target: any, key: string, value: any): null {
+function decorator(target: any, key: string, value: any, options: ReviserDecoratorOptions): null {
+
+  if (!options.required && !isRequired(value)) {
+    return null;
+  }
+
+  const type: string = getPrimitiveType(value);
+
+  if (type === PrimitiveTypes.Undefined || type === PrimitiveTypes.Null) {
+    target[key] = "";
+    return null;
+  }
+
   target[key] = String(value).valueOf();
   return null;
 }