+++ src/impl/translators/to-uint32.ts
@@ -5,10 +5,15 @@
  */
 
 import { factory } from "@/decorator-factory";
-import { PropertyDecorator, ReviserMessage } from "@inter/decorator";
+import { PropertyDecorator, ReviserMessage, ReviserDecoratorOptions } from "@inter/decorator";
 import { IEEE754Limits } from "@/constants";
+import { isRequired } from "@impl/utils";
 
-function decorator(target: any, key: string, value: any): null {
+function decorator(target: any, key: string, value: any, options: ReviserDecoratorOptions): null {
+  if (!options.required && !isRequired(value)) {
+    return null;
+  }
+
   let num: number = Number(value).valueOf() || 0;
 
   if (num !== Infinity && num !== -Infinity && num % 1 !== 0) {