+++ src/impl/types/type-email.ts
@@ -5,8 +5,8 @@
  */
 
 import { factory } from "@/decorator-factory";
-import { PropertyDecorator, ReviserDecoratorReturns } from "@inter/decorator";
-import { PrimitiveTypes, getPrimitiveType } from "@impl/utils";
+import { PropertyDecorator, ReviserDecoratorReturns, ReviserDecoratorOptions } from "@inter/decorator";
+import { PrimitiveTypes, getPrimitiveType, isRequired } from "@impl/utils";
 import { parse, getTemplate } from "@impl/message";
 
 const Pattern = /^([\w-_]+(?:\.[\w-_]+)*)@((?:[a-z0-9]+(?:-[a-zA-Z0-9]+)*)+\.[a-z]{2,6})$/;
@@ -29,7 +29,11 @@ function TypeEmail(template?: string | Templates): PropertyDecorator {
     pattern: getTemplate(Def.pattern, "pattern", template)
   };
 
-  function decorator(target: any, key: string, value: any): ReviserDecoratorReturns<{}> {
+  function decorator(target: any, key: string, value: any, options: ReviserDecoratorOptions): ReviserDecoratorReturns<{}> {
+    if (!options.required && !isRequired(value)) {
+      return null;
+    }
+
     const type: string = getPrimitiveType(value);
 
     if (type !== PrimitiveTypes.String) {