+++ test/decorator-type.test.ts
@@ -8,7 +8,7 @@ import "mocha";
 import { expect, assert } from "chai";
 import {
   Reviser, TypeBoolean, TypeDouble, TypeEmail, TypeFloat, TypeInt32,
-  TypeInt64, TypePhone, TypeString, TypeStruct, TypeUnInt32, TypeUnInt64
+  TypeInt64, TypePhone, TypeString, TypeStruct, TypeUnInt32, TypeUnInt64, Required,
 } from "@/index";
 
 // TODO
@@ -17,9 +17,21 @@ import {
 describe("Data type decorators", function() {
   // TypeBoolean
   describe("@TypeBoolean", function() {
+    it("Should ignore non-required property while pass a non-required value", function() {
+      class M extends Reviser {
+        @TypeBoolean()
+        p = true;
+      };
+
+      const m = new M();
+      const message = m.map({});
+      expect(message).to.be.a('null');
+    });
+
     it("Should return string message while map a non-boolean data", function() {
       class M extends Reviser {
         @TypeBoolean()
+        @Required()
         p = true;
       };
 
@@ -31,6 +43,7 @@ describe("Data type decorators", function() {
     it("Should return null while map a boolean data", function() {
       class M extends Reviser {
         @TypeBoolean()
+        @Required()
         p = true;
       };
 
@@ -44,6 +57,7 @@ describe("Data type decorators", function() {
 
       class M extends Reviser {
         @TypeBoolean(customMesage)
+        @Required()
         p = true;
       };
 
@@ -58,6 +72,7 @@ describe("Data type decorators", function() {
     it("Should return string message while map a non-double data", function() {
       class M extends Reviser {
         @TypeDouble()
+        @Required()
         p = 1;
       };
 
@@ -69,6 +84,7 @@ describe("Data type decorators", function() {
     it("Should return null while map a double data", function() {
       class M extends Reviser {
         @TypeDouble()
+        @Required()
         p = 1;
       };
 
@@ -82,6 +98,7 @@ describe("Data type decorators", function() {
 
       class M extends Reviser {
         @TypeDouble(customMesage)
+        @Required()
         p = 1;
       };
 
@@ -96,6 +113,7 @@ describe("Data type decorators", function() {
     it("Should return string message while map a non-email data", function() {
       class M extends Reviser {
         @TypeEmail()
+        @Required()
         p = "";
       };
 
@@ -107,6 +125,7 @@ describe("Data type decorators", function() {
     it("Should return null while map a email data", function() {
       class M extends Reviser {
         @TypeEmail()
+        @Required()
         p = "";
       };
 
@@ -120,6 +139,7 @@ describe("Data type decorators", function() {
 
       class M extends Reviser {
         @TypeEmail(customMesage)
+        @Required()
         p = 1;
       };
 
@@ -134,6 +154,7 @@ describe("Data type decorators", function() {
     it("Should return string message while map a non-float data", function() {
       class M extends Reviser {
         @TypeFloat()
+        @Required()
         p = 1.2;
       };
 
@@ -145,6 +166,7 @@ describe("Data type decorators", function() {
     it("Should return null while map a float data", function() {
       class M extends Reviser {
         @TypeFloat()
+        @Required()
         p = 1.2;
       };
 
@@ -158,6 +180,7 @@ describe("Data type decorators", function() {
 
       class M extends Reviser {
         @TypeFloat(customMesage)
+        @Required()
         p = 1;
       };
 
@@ -172,6 +195,7 @@ describe("Data type decorators", function() {
     it("Should return string message while map a non-int32 data", function() {
       class M extends Reviser {
         @TypeInt32()
+        @Required()
         p = 1;
       };
 
@@ -183,6 +207,7 @@ describe("Data type decorators", function() {
     it("Should return null while map a int32 data", function() {
       class M extends Reviser {
         @TypeInt32()
+        @Required()
         p = 1;
       };
 
@@ -196,6 +221,7 @@ describe("Data type decorators", function() {
 
       class M extends Reviser {
         @TypeInt32(customMesage)
+        @Required()
         p = 1;
       };
 
@@ -210,6 +236,7 @@ describe("Data type decorators", function() {
     it("Should return string message while map a non-int64 data", function() {
       class M extends Reviser {
         @TypeInt64()
+        @Required()
         p = 1;
       };
 
@@ -221,6 +248,7 @@ describe("Data type decorators", function() {
     it("Should return null while map a int64 data", function() {
       class M extends Reviser {
         @TypeInt64()
+        @Required()
         p = 1;
       };
 
@@ -234,6 +262,7 @@ describe("Data type decorators", function() {
 
       class M extends Reviser {
         @TypeInt64(customMesage)
+        @Required()
         p = 1;
       };
 
@@ -248,6 +277,7 @@ describe("Data type decorators", function() {
     it("Should return string message while map a non-phone data", function() {
       class M extends Reviser {
         @TypePhone()
+        @Required()
         p = "";
       };
 
@@ -259,6 +289,7 @@ describe("Data type decorators", function() {
     it("Should return null while map a phone data", function() {
       class M extends Reviser {
         @TypePhone()
+        @Required()
         p = 1;
       };
 
@@ -272,6 +303,7 @@ describe("Data type decorators", function() {
 
       class M extends Reviser {
         @TypePhone(customMesage)
+        @Required()
         p = 1;
       };
 
@@ -286,6 +318,7 @@ describe("Data type decorators", function() {
     it("Should return string message while map a non-string data", function() {
       class M extends Reviser {
         @TypeString()
+        @Required()
         p = "";
       };
 
@@ -297,6 +330,7 @@ describe("Data type decorators", function() {
     it("Should return null while map a string data", function() {
       class M extends Reviser {
         @TypeString()
+        @Required()
         p = 1;
       };
 
@@ -310,6 +344,7 @@ describe("Data type decorators", function() {
 
       class M extends Reviser {
         @TypeString(customMesage)
+        @Required()
         p = 1;
       };
 
@@ -324,9 +359,11 @@ describe("Data type decorators", function() {
     it("Should return string message while map a non-structure class", function() {
       class B extends Reviser {
         @TypeInt32()
+        @Required()
         foo = 1;
 
         @TypeString()
+        @Required()
         bar = "";
       };
 
@@ -345,9 +382,11 @@ describe("Data type decorators", function() {
     it("Should return null while map a structure class", function() {
       class B extends Reviser {
         @TypeInt32()
+        @Required()
         foo = 1;
 
         @TypeString()
+        @Required()
         bar = "";
       };
 
@@ -372,6 +411,7 @@ describe("Data type decorators", function() {
     it("Should return string message while map a non-uint32 data", function() {
       class M extends Reviser {
         @TypeUnInt32()
+        @Required()
         p = 1;
       };
 
@@ -383,6 +423,7 @@ describe("Data type decorators", function() {
     it("Should return null while map a uint32 data", function() {
       class M extends Reviser {
         @TypeUnInt32()
+        @Required()
         p = 1;
       };
 
@@ -396,6 +437,7 @@ describe("Data type decorators", function() {
 
       class M extends Reviser {
         @TypeUnInt32(customMesage)
+        @Required()
         p = 1;
       };